requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web Scraping
selenium>=4.10.0
webdriver-manager>=4.0.0

# Statistical Analysis
statsmodels>=0.14.0
linearmodels>=5.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: RAG with pgvector
# pgvector>=0.2.0
# sentence-transformers>=2.2.0
